fix(earth) edit folder
