+++ src/components/profile/ProfileMobile.tsx
@@ -1,4 +1,6 @@
 import React, { useState } from "react";
+import EditButton from "../button/EditButton";
+
 
 const MobileProfilePage: React.FC = () => {
   const [isOrderhistoryVisible, setIsOrderhistorVisible] = useState(false);
@@ -12,7 +14,7 @@ const MobileProfilePage: React.FC = () => {
   const togglePayments = () => setIsPaymentsVisible(!isPaymentsVisible);
 
   return (
-    <div className='flex min-h-screen flex-col items-center p-4 md:hidden'>
+    <div className='flex h-full flex-col items-center py-4 md:hidden'>
       <div className='w-full max-w-xl bg-white p-6 shadow-md'>
         {/* Account Section */}
         <div className='flex items-center justify-between border-b py-2'>
@@ -80,22 +82,7 @@ const MobileProfilePage: React.FC = () => {
                 </p>
               </div>
               <div>
-                <button>
-                  <svg
-                    xmlns='http://www.w3.org/2000/svg'
-                    fill='none'
-                    viewBox='0 0 24 24'
-                    strokeWidth='1.5'
-                    stroke='currentColor'
-                    className='h-6 w-6'
-                  >
-                    <path
-                      strokeLinecap='round'
-                      strokeLinejoin='round'
-                      d='m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10'
-                    />
-                  </svg>
-                </button>
+                <EditButton  />
               </div>
             </div>
           )}
